requirements.txt
langchain==0.0.306
python-dotenv==1.0.0
openai==0.28.1
matplotlib==3.8.0
plotly==5.17.0
pandas==2.1.1
scipy==1.11.3
scikit-learn==1.3.1